+++ NetFabric.Hyperlinq.Benchmarks/First.Benchmarks.cs
@@ -33,12 +33,12 @@ namespace NetFabric.Hyperlinq.Benchmarks
 
         [BenchmarkCategory("Enumerable_Reference")]
         [Benchmark(Baseline = true)]
-        public int Linq_Enumerable_Reference() => 
+        public long Linq_Enumerable_Reference() => 
             System.Linq.Enumerable.First(enumerableReference);
 
         [BenchmarkCategory("Enumerable_Value")]
         [Benchmark(Baseline = true)]
-        public int Linq_Enumerable_Value() => 
+        public long Linq_Enumerable_Value() => 
             System.Linq.Enumerable.First(enumerableValue);
 
         [BenchmarkCategory("Array")]
@@ -53,7 +53,7 @@ namespace NetFabric.Hyperlinq.Benchmarks
 
         [BenchmarkCategory("Range")]
         [Benchmark]
-        public int Hyperlinq_Range() =>
+        public long Hyperlinq_Range() =>
             hyperlinqRange.First();
 
         [BenchmarkCategory("Queue")]
@@ -73,12 +73,12 @@ namespace NetFabric.Hyperlinq.Benchmarks
 
         [BenchmarkCategory("Enumerable_Reference")]
         [Benchmark]
-        public int Hyperlinq_Enumerable_Reference() => 
+        public long Hyperlinq_Enumerable_Reference() => 
             enumerableReference.First();
 
         [BenchmarkCategory("Enumerable_Value")]
         [Benchmark]
-        public int Hyperlinq_Enumerable_Value() => 
-            enumerableValue.First<TestEnumerable.Enumerable, TestEnumerable.Enumerable.Enumerator, int>();
+        public long Hyperlinq_Enumerable_Value() => 
+            enumerableValue.First<TestEnumerable.Enumerable, TestEnumerable.Enumerable.Enumerator, long>();
     }
 }
