+++ NetFabric.Hyperlinq.UnitTests/FluentAssertions/Extensions/ValueReadOnlyListExtensions.cs
@@ -12,7 +12,7 @@ namespace NetFabric.Hyperlinq.UnitTests
             where TEnumeratorFirst : struct, IValueEnumerator<TFirst>
         {
             if (first.Count != second.Count)
-                return Math.Min(first.Count, second.Count);
+                return (int)Math.Min(first.Count, second.Count);
 
             for (var index = 0; index < first.Count; index++)
             {
